rename topics of base nodes

+++ setup/robot/tmux/champi_start.bash
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-sudo systemctl stop champi.service
-sudo systemctl start champi.service
+if ! tmux has-session -t champi_session 2>/dev/null; then
+  sudo systemctl stop champi.service
+  sudo systemctl start champi.service
+fi
+
 tmux attach -t champi_session
